\name{EmptyModel}
\alias{EmptyModel}
\title{Empty Planet Model}
\description{Create an empty planet model with defined, named elements
  including NaN or length 0 values. }
\usage{
EmptyModel()
}

\value{Planet model containing no information.}

\author{ Jake Anderson }

\examples{
EmptyModel()
}
\keyword{ misc }
