\name{EmptyModel}
\alias{EmptyModel}
\title{Empty Planet Model}
\description{Create an empty planet model with defined, named elements
  including NaN or length 0 values. }
\usage{
EmptyModel()
}
\details{
}
\value{Planet model containing no information.}
\references{
}
\author{ Jake Anderson }
\note{
}


\seealso{
}
\examples{
EmptyModel()
}
\keyword{ misc }
